--- src/plccng/parse_spec/parse_spec/__init__.py
@@ -1,5 +1,0 @@
-from ..structs import CodeFragment, Line, SemanticSpec, TargetLocator
-from .parse_syntactic_spec import parse_syntactic_spec
-from .parse_lexical_spec import parse_lexical_spec
-
-from .parse_semantic_spec import parse_semantic_spec
