--- src/plccng/parse_spec/parse_spec/parse_semantic_spec/__init__.py
@@ -1,4 +1,0 @@
-from ...structs import CodeFragment, SemanticSpec, TargetLocator
-from .parse_semantic_spec import parse_semantic_spec
-from .parse_code_fragments import parse_code_fragments
-from .parse_target_locator import parse_target_locator
