--- src/plccng/parse_spec/parse_spec/parse_syntactic_spec/__init__.py
@@ -1,5 +1,0 @@
-from ...structs import CapturingSymbol, CapturingTerminal, LhsNonTerminal, RepeatingSyntacticRule, RhsNonTerminal, StandardSyntacticRule, Symbol, SyntacticRule, SyntacticSpec, Terminal
-from .parse_syntactic_spec import parse_syntactic_spec
-from ...errors import (
-    MalformedBNFError,
-)
