--- src/plccng/parse_spec/validate_spec/validate_lexical_spec/__init__.py
@@ -1,8 +1,0 @@
-from ...structs import Block, Divider, Include, LexicalRule, LexicalSpec, Line
-from .validate_lexical_spec import validate_lexical_spec
-
-from ...parse_spec.parse_lexical_spec import parse_lexical_spec
-from ...parse_rough.parse_blocks import parse_blocks, UnclosedBlockError
-from ...parse_rough.parse_includes import parse_includes
-from ...parse_rough.parse_dividers import parse_dividers
-from ...parse_rough.parse_rough import from_string
