--- src/plccng/parse_spec/validate_spec/validate_semantic_spec/__init__.py
@@ -1,2 +1,0 @@
-from ...structs import Line
-from .validate_semantic_spec import validate_semantic_spec
