--- src/plccng/parse_spec/validate_spec/validate_syntactic_spec/check_ll1/__init__.py
@@ -1,1 +1,0 @@
-from .check_ll1 import check_ll1
